pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-search-notes"
version = "0.1.0"
description = "Semantic search system for personal notes using vector embeddings"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Dhavan", email = "codingquark@gluon.lan"}
]
keywords = ["semantic-search", "notes", "embeddings", "qdrant", "flask"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.21.0",
    "sentence-transformers>=2.2.0",
    "qdrant-client>=1.6.0",
    "flask>=2.3.0",
    "gunicorn>=21.0.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
gpu = [
    "faiss-gpu>=1.7.0",
]
cpu = [
    "faiss-cpu>=1.7.0",
]

[project.scripts]
semantic-notes-index = "semantic_notes.cli:index_command"
semantic-notes-serve = "semantic_notes.cli:serve_command"

[project.urls]
"Homepage" = "https://github.com/codingquark/semantic-search-notes"
"Bug Reports" = "https://github.com/codingquark/semantic-search-notes/issues"
"Source" = "https://github.com/codingquark/semantic-search-notes"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/semantic_notes --cov-report=term-missing"